Fetch all commits before finding first commit
GitHub's action/checkout action creates a shallow clone,
`initial_commit` returns the most recent commit SHA instead of the
first. Show teach it to detect a shallow clone and fetch all commits, so
that it can then fetch the earliest SHA. This only happens for the first
release in a Changelog.