Use full path to call `mknotes`
v0.1.2 adopted the [composite action howto]'s pattern of adding the
action path to `GITHUB_PATH`. But when I went to make a release of a
project today, it was unable to find `mknotes`. So use the full path to
`mknotes`, intead. Not sure why this works, but the [syntax docs]
demonstrate using the full path, and that fixed the problem.

While at it, update the test workflow to always test the
currently-cloned branch. It did not catch this issue, but should allow
future changes to immediately appear in the test results, instead of
always testing main.

  [composite action howto]: https://docs.github.com/en/actions/sharing-automations/creating-actions/creating-a-composite-action
  [syntax docs]: https://docs.github.com/en/actions/sharing-automations/creating-actions/metadata-syntax-for-github-actions#runssteps